--- src/client/login/components/styled.tsx
@@ -1,13 +1,0 @@
-import GoogleButton from 'react-google-button';
-import styled from 'styled-components';
-
-export const LoginContainer = styled.div`
-  padding: 15px;
-  display: flex;
-  width: 100%;
-  height: 50vh;
-`;
-
-export const StyledGoogleButton = styled(GoogleButton)`
-  margin: auto;
-`;
