+++ src/client/team/components/login.tsx
@@ -1,5 +1,6 @@
 import React from 'react';
 import { LoginContainer, StyledGoogleButton } from './styled';
+
 interface Props {
   signIn: () => void;
 }
