+++ src/client/team/components/styled.tsx
@@ -4,6 +4,7 @@ import {
   FormControlLabel,
   IconButton,
 } from '@mui/material';
+import GoogleButton from 'react-google-button';
 import styled from 'styled-components';
 
 export const TeamPageContainer = styled.div`
@@ -148,4 +149,15 @@ export const Error = styled.p`
   color: #ff3e51;
   font-size: 14px;
   margin-top: 4px;
+`;
+
+export const LoginContainer = styled.div`
+  padding: 15px;
+  display: flex;
+  width: 100%;
+  height: 50vh;
+`;
+
+export const StyledGoogleButton = styled(GoogleButton)`
+  margin: auto;
 `;
